Safari-611.1.10/LayoutTests/css3/shapes/shape-outside/values/shape-image-threshold-002-expected.txt


FAIL -0.5 - inline - invalid assert_equals: expected (object) null but got (string) ""
FAIL -100 - inline - invalid assert_equals: expected (object) null but got (string) ""
FAIL nonNumber - inline - invalid assert_equals: expected (object) null but got (string) ""
FAIL 3/4 - inline - invalid assert_equals: expected (object) null but got (string) ""
FAIL 20% - inline - invalid assert_equals: expected (object) null but got (string) ""
